+++ JudoPayDotNet/Models/ApplePayPaymentModel.cs
@@ -1,18 +1,14 @@
 ﻿using System.Runtime.Serialization;
 
 // ReSharper disable ClassNeverInstantiated.Global
-
 namespace JudoPayDotNet.Models
 {
     [DataContract]
-    public class PKPaymentModel : PaymentModel
+    public class ApplePayPaymentModel : PaymentModel
     {
         /// <summary>
-        /// Gets or sets the apple pay token.
+        /// Apple Pay transaction details.
         /// </summary>
-        /// <value>
-        /// The apple pay token.
-        /// </value>
         [DataMember(IsRequired = true)]
         public PKPaymentInnerModel PkPayment { get; set; }
     }